refactor: remove abstrast document
