+++ pyproject.toml
@@ -75,7 +75,9 @@ skip_glob= ['docarray/proto/pb2/*', 'docarray/proto/pb/*']
 
 [tool.ruff]
 exclude = ['docarray/proto/pb2/*', 'docarray/proto/pb/*', 'docs/*']
-
+ignore = [
+    "E501",  # line too long, handled by black
+]
 [tool.pytest.ini_options]
 markers = [
     "slow: marks tests as slow (deselect with '-m \"not slow\"')",
